Remove default values for key parameters
Having default values for the key parameters in dir-contents caused
errors on the main function, since it called dir-contents with values
not specified by the user as "nil", which was the wrong type. Now the
default values are set in the body of the funtion if the key parameter
is nil.